+++ mwmbl/indexer/bootstrap.sh
