+++ mwmbl/tinysearchengine/static/index.js
