+++ mwmbl/tinysearchengine/static/plugin.xml
